Features 02: Comments
In this feature, users can post comments on the image posts

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -1,8 +1,10 @@
 package ImageHoster.controller;
 
+import ImageHoster.model.Comment;
 import ImageHoster.model.Image;
 import ImageHoster.model.Tag;
 import ImageHoster.model.User;
+import ImageHoster.service.CommentService;
 import ImageHoster.service.ImageService;
 import ImageHoster.service.TagService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,6 +29,11 @@ public class ImageController {
     @Autowired
     private TagService tagService;
 
+    @Autowired
+    private CommentService commentService;
+
+    private String notPermitted = "";
+
     //This method displays all the images in the user home page after successful login
     @RequestMapping("images")
     public String getUserImages(Model model) {
@@ -45,11 +52,25 @@ public class ImageController {
     //Also now you need to add the tags of an image in the Model type object
     //Here a list of tags is added in the Model type object
     //this list is then sent to 'images/image.html' file and the tags are displayed
-    @RequestMapping("/images/{title}")
-    public String showImage(@PathVariable("title") String title, Model model) {
-        Image image = imageService.getImageByTitle(title);
+    @RequestMapping("/images/{imageId}/{title}")
+    public String showImage(@PathVariable("imageId") Integer imageId, Model model) {
+        String error;
+        Image image = imageService.getImage(imageId);
+        if (notPermitted != "") {
+
+            if (notPermitted.equalsIgnoreCase("Edit")) {
+                error = "Only the owner of the image can edit the image";
+                model.addAttribute("editError", error);
+            } else {
+                error = "Only the owner of the image can delete the image";
+                model.addAttribute("deleteError", error);
+            }
+            notPermitted = "";
+        }
         model.addAttribute("image", image);
         model.addAttribute("tags", image.getTags());
+        List<Comment> commentList = commentService.getAllComments(image.getId(), image.getTitle());
+        model.addAttribute("comments", commentList);
         return "images/image";
     }
 
@@ -92,13 +113,20 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
         Image image = imageService.getImage(imageId);
-
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        User user = (User) session.getAttribute("loggeduser");
+        int uId = user.getId();
+        int iId = image.getUser().getId();
+        if (uId == iId) {
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        } else {
+            notPermitted = "Edit";
+            return "redirect:/images/" + image.getId() + "/" + image.getTitle();
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -113,7 +141,8 @@ public class ImageController {
     //The method also receives tags parameter which is a string of all the tags separated by a comma using the annotation @RequestParam
     //The method converts the string to a list of all the tags using findOrCreateTags() method and sets the tags attribute of an image as a list of all the tags
     @RequestMapping(value = "/editImage", method = RequestMethod.PUT)
-    public String editImageSubmit(@RequestParam("file") MultipartFile file, @RequestParam("imageId") Integer imageId, @RequestParam("tags") String tags, Image updatedImage, HttpSession session) throws IOException {
+    public String editImageSubmit(@RequestParam("file") MultipartFile file, @RequestParam("imageId") Integer
+            imageId, @RequestParam("tags") String tags, Image updatedImage, HttpSession session) throws IOException {
 
         Image image = imageService.getImage(imageId);
         String updatedImageData = convertUploadedFileToBase64(file);
@@ -140,11 +169,30 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, Model
+            model) {
+        String error = "Only the owner of the image can delete the image";
+        User user = (User) session.getAttribute("loggeduser");
+        Image image = imageService.getImage(imageId);
+        int uId = user.getId();
+        int iId = image.getUser().getId();
+        if (uId == iId) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        } else {
+            notPermitted = "Delete";
+            return "redirect:/images/" + image.getId() + "/" + image.getTitle();
+        }
     }
 
+    //@RequestMapping(value = "/image/{image.id}/{image.title}/comments", method = RequestMethod.POST)
+    //public String postComment(Comment comment, HttpSession session) throws IOException {
+
+    // User user = (User) session.getAttribute("loggeduser");
+    // comment.setUser(user);
+    // imageService.uploadImage(newImage);
+    // return "redirect://image/{image.id}/{image.title}";
+    // }
 
     //This method converts the image to Base64 format
     private String convertUploadedFileToBase64(MultipartFile file) throws IOException {
